+++ R/track_transform_crs.R
@@ -20,12 +20,41 @@ transform_coords <- function(x, ...) {
 #' get_crs(deer)
 #'
 #' # project to geographical coordinates (note the CRS is taken automatically from the object deer).
-#' d1 <- transform_coords(deer, crs = 4326)
+#' d1 <- transform_coords(deer, crs_to = 4326)
 
 transform_coords.track_xy <- function(x, crs_to, crs_from, ...) {
+
   if (missing(crs_from)) {
     if (!is.null(attributes(x)$crs)) {
       crs_from <- attributes(x)$crs
+    }
+  }
+
+  # crs_to
+  if (!is.na(crs_to)) {
+    if (is(crs_to, "CRS")) {
+      .Deprecated("It looks like you used `CRS()` to create the crs_to,
+                  please use the ESPG directly.")
+      crs <- sf::st_crs(crs_to)
+    } else {
+      crs <- sf::st_crs(crs_to)
+      if (is.na(crs_to)) {
+        warning("`crs_to` invalid.")
+      }
+    }
+  }
+
+  # crs_from
+  if (!is.na(crs_from)) {
+    if (is(crs_from, "CRS")) {
+      .Deprecated("It looks like you used `CRS()` to create the crs_from,
+                  please use the ESPG directly.")
+      crs <- sf::st_crs(crs_from)
+    } else {
+      crs <- sf::st_crs(crs_from)
+      if (is.na(crs_from)) {
+        warning("`crs_from` invalid.")
+      }
     }
   }